pyproject.toml
[project]
name = "p10-dsml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "easyocr>=1.7.2",
    "faiss-cpu==1.10.0",
    "ipykernel>=7.1.0",
    "langchain==0.3.23",
    "langchain-huggingface>=0.3.1",
    "langchain-mistralai>=0.2.12",
    "logfire>=4.22.0",
    "mistralai==0.4.2",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "pillow>=9.0.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.22.0",
    "pypdf2==3.0.1",
    "python-docx==1.1.2",
    "python-dotenv==1.1.0",
    "ragas==0.1.4",
    "requests==2.32.3",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "streamlit==1.44.1",
    "streamlit-feedback==0.1.4",
    "tqdm>=4.67.1",
]
